make it work right: Extern edition
So having a static copy of everything floating around is a bad idea.

I've fixed that here: modem_state now has its own compilation unit that
branches out into the other areas. This could theoretically make it so
that the various modes are held in modem_state, but I wanted to make the
cross-class stuff as minimal as possible.